fix focus point with double rounding error (#3247)

+++ src/Sulu/Bundle/MediaBundle/Media/ImageConverter/Focus/Focus.php
@@ -84,7 +84,7 @@ class Focus implements FocusInterface
             }
         } else {
             $width = $imageSize->getWidth();
-            $height = $width / $targetRatio;
+            $height = (int) ($width / $targetRatio);
 
             $cropX = 0;
